test(orb): decouple CCG bootstrap helpers from runtime
