+++ orb/engine/package.json
@@ -5,7 +5,7 @@
     "main": "workflow-assistant.js",
     "scripts": {
         "test": "node tests/music-composition-studio.test.js && node --test tests/campaign-creative-executor.test.js",
-        "lint": "node --check music-composition-studio/pipeline.js && node --check music-composition-studio/adapters.js && node --check music-composition-studio/error-handler.js && node --check music-composition-studio/lib/canonical.js && node --check music-composition-studio/lib/compatibility.js && node --check music-composition-studio/lib/contracts.js && node --check music-composition-studio/lib/invalidation.js && node --check music-composition-studio/lib/ledger.js && node --check music-composition-studio/lib/providers.js && node --check music-composition-studio/lib/schema-definitions.js && node --check music-composition-studio/services/audio-service.js && node --check campaign-creative-executor/contracts.js && node --check campaign-creative-executor/storage.js && node --check campaign-creative-executor/store.js && node --check campaign-creative-executor/adapters/renderer.js && node --check campaign-creative-executor/adapters/mock.js && node --check campaign-creative-executor/adapters/openai-images.js && node --check campaign-creative-executor/adapters/storage.js && node --check campaign-creative-executor/registry.js && node --check campaign-creative-executor/executor.js && node --check campaign-creative-executor/server.js && node --check campaign-creative-executor/index.js && node --check scripts/build-music-composition-studio.js && node --check scripts/dry-run-music-composition-studio.js && node --check scripts/generate-music-composition-studio-fixtures.js && node --check scripts/generate-music-composition-studio-schemas.js && node --check scripts/validate-music-composition-studio.js && node --check scripts/build-campaign-creative-creator-continuous.js && node --check scripts/preload-n8n-error-workflow-bootstrap.js && node --check scripts/prepare-campaign-creative-creator-live-import.js && node --check scripts/validate-campaign-creative-creator-continuous.js && node --check scripts/build-campaign-creative-creator-organizer.js && node --check tests/campaign-creative-creator-continuous.test.js && node --check tests/campaign-creative-creator-error-trigger-runtime.test.js && node --check tests/campaign-creative-creator-live-import.test.js && node --check tests/campaign-creative-executor.test.js && bash -n scripts/start-n8n-runtime.sh && bash -n scripts/validate-music-composition-studio-migration.sh && bash -n scripts/validate-music-composition-studio-n8n-import.sh",
+        "lint": "node --check music-composition-studio/pipeline.js && node --check music-composition-studio/adapters.js && node --check music-composition-studio/error-handler.js && node --check music-composition-studio/lib/canonical.js && node --check music-composition-studio/lib/compatibility.js && node --check music-composition-studio/lib/contracts.js && node --check music-composition-studio/lib/invalidation.js && node --check music-composition-studio/lib/ledger.js && node --check music-composition-studio/lib/providers.js && node --check music-composition-studio/lib/schema-definitions.js && node --check music-composition-studio/services/audio-service.js && node --check campaign-creative-executor/contracts.js && node --check campaign-creative-executor/storage.js && node --check campaign-creative-executor/store.js && node --check campaign-creative-executor/adapters/renderer.js && node --check campaign-creative-executor/adapters/mock.js && node --check campaign-creative-executor/adapters/openai-images.js && node --check campaign-creative-executor/adapters/storage.js && node --check campaign-creative-executor/registry.js && node --check campaign-creative-executor/executor.js && node --check campaign-creative-executor/server.js && node --check campaign-creative-executor/index.js && node --check scripts/build-music-composition-studio.js && node --check scripts/dry-run-music-composition-studio.js && node --check scripts/generate-music-composition-studio-fixtures.js && node --check scripts/generate-music-composition-studio-schemas.js && node --check scripts/validate-music-composition-studio.js && node --check scripts/build-campaign-creative-creator-continuous.js && node --check scripts/n8n-error-workflow-bootstrap-core.js && node --check scripts/preload-n8n-error-workflow-bootstrap.js && node --check scripts/prepare-campaign-creative-creator-live-import.js && node --check scripts/validate-campaign-creative-creator-continuous.js && node --check scripts/build-campaign-creative-creator-organizer.js && node --check tests/campaign-creative-creator-continuous.test.js && node --check tests/campaign-creative-creator-error-trigger-runtime.test.js && node --check tests/campaign-creative-creator-live-import.test.js && node --check tests/campaign-creative-executor.test.js && bash -n scripts/start-n8n-runtime.sh && bash -n scripts/validate-music-composition-studio-migration.sh && bash -n scripts/validate-music-composition-studio-n8n-import.sh",
         "workflow:music:schemas": "node scripts/generate-music-composition-studio-schemas.js",
         "workflow:music:fixtures": "node scripts/generate-music-composition-studio-fixtures.js",
         "workflow:music:build": "node scripts/generate-music-composition-studio-schemas.js && node scripts/generate-music-composition-studio-fixtures.js && node scripts/build-music-composition-studio.js",
@@ -26,6 +26,7 @@
         "workflow:campaign-creative:continuous:build": "node scripts/build-campaign-creative-creator-continuous.js",
         "workflow:campaign-creative:continuous:validate": "node scripts/validate-campaign-creative-creator-continuous.js",
         "workflow:campaign-creative:continuous:test": "node tests/campaign-creative-creator-continuous.test.js",
+        "workflow:campaign-creative:runtime:test": "node --test tests/campaign-creative-creator*.test.js",
         "workflow:campaign-creative:error-trigger:test": "node --test tests/campaign-creative-creator-error-trigger-runtime.test.js",
         "workflow:campaign-creative:live-import:test": "node --test tests/campaign-creative-creator-live-import.test.js",
         "workflow:campaign-creative:executor:test": "node --test tests/campaign-creative-executor.test.js",
